+++ CHANGELOG.md
@@ -52,7 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `postalCode` from storage to Session context.
 - Updates all tokens naming and simplifies the global styles.
 - Changes `theme.scss` file to `global/tokens.scss`.
-- `OutOfStock` add style and success message.
+- `OutOfStock` style and success message.
 
 ### Deprecated